Rename Header → CollectionHeader
